// frontend/src/api/generated/model/ndmsNativeOwnershipLifecycle.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * keen-pbr API
 * REST API for the keen-pbr policy-based routing daemon.
 * OpenAPI spec version: 3.0.0
 */

export type NdmsNativeOwnershipLifecycle = typeof NdmsNativeOwnershipLifecycle[keyof typeof NdmsNativeOwnershipLifecycle];


export const NdmsNativeOwnershipLifecycle = {
  active_running_only: 'active_running_only',
  active_save_acknowledged_unverified: 'active_save_acknowledged_unverified',
  deleted_save_acknowledged_unverified: 'deleted_save_acknowledged_unverified',
} as const;